@@ -1,103 +1,73 @@
 {
   "name": "web-check",
-  "version": "1.1.2",
-  "private": false,
-  "description": "All-in-one OSINT tool for analyzing any website",
-  "repository": "github:lissy93/web-check",
+  "type": "module",
+  "version": "2.0.0",
   "homepage": "https://web-check.xyz",
-  "license": "MIT",
-  "author": {
-    "name": "Alicia Sykes",
-    "email": "alicia@omg.lol"
-  },
-  "funding": {
-    "type": "github",
-    "url": "https://github.com/sponsors/Lissy93"
-  },
   "scripts": {
-    "dev": "netlify dev",
-    "serve": "node server",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "node server",
+    "start-pm": "pm2 start server.js -i max",
+    "build": "astro check && astro build",
     "dev:vercel": "PLATFORM='vercel' npx vercel dev",
     "dev:netlify": "PLATFORM='netlify' npx netlify dev",
-    "dev:node": "npx concurrently --names \"frontend,backend\" \"REACT_APP_API_ENDPOINT='http://localhost:3001/api' npx react-scripts start\" \"PLATFORM='node' DISABLE_GUI='true' PORT='3001' API_CORS_ORIGIN='*' npx nodemon server\""
+    "dev:api": "DISABLE_GUI='true' PORT='3001' nodemon server",
+    "dev:astro": "PUBLIC_API_ENDPOINT=http://localhost:3001/api astro dev",
+    "dev": "concurrently -c magenta,cyan -n backend,frontend 'yarn dev:api' 'yarn dev:astro'"
   },
   "dependencies": {
-    "@netlify/functions": "^1.6.0",
-    "@sentry/react": "^7.60.0",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.5.3",
-    "@types/node": "^20.4.4",
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/react-simple-maps": "^3.0.0",
-    "@types/styled-components": "^5.1.26",
-    "axios": "^1.4.0",
+    "@astrojs/check": "^0.5.10",
+    "@astrojs/react": "^3.3.2",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-brands-svg-icons": "^6.5.2",
+    "@fortawesome/free-regular-svg-icons": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@fortawesome/svelte-fontawesome": "^0.2.2",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.0",
+    "astro": "^4.7.1",
+    "axios": "^1.4.8",
     "cheerio": "^1.0.0-rc.12",
     "chrome-aws-lambda": "^10.1.0",
     "chromium": "^3.0.3",
     "connect-history-api-fallback": "^2.0.0",
     "cors": "^2.8.5",
     "csv-parser": "^3.0.0",
-    "dotenv": "^16.3.1",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
     "express-rate-limit": "^7.2.0",
-    "flatted": "^3.2.7",
-    "follow-redirects": "^1.15.2",
-    "got": "^13.0.0",
-    "jest-styled-components": "^7.1.1",
-    "netlify-cli": "^15.9.1",
-    "perf_hooks": "^0.0.1",
+    "framer-motion": "^11.2.6",
+    "got": "^14.2.1",
+    "pm2": "^5.3.1",
     "psl": "^1.9.0",
-    "puppeteer": "^20.9.0",
-    "puppeteer-core": "^21.0.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "puppeteer": "^22.8.0",
+    "puppeteer-core": "^22.8.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-masonry-css": "^1.0.16",
-    "react-router-dom": "^6.14.2",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.23.0",
     "react-simple-maps": "^3.0.0",
-    "react-toastify": "^9.1.3",
-    "recharts": "^2.7.3",
-    "styled-components": "^6.0.5",
+    "react-toastify": "^10.0.5",
+    "recharts": "^2.12.6",
+    "svelte": "^4.2.17",
     "traceroute": "^1.0.0",
-    "typescript": "^5.1.6",
-    "unzipper": "^0.10.14",
+    "typescript": "^5.4.5",
+    "unzipper": "^0.11.5",
+    "url-parse": "^1.5.10",
     "wappalyzer": "^6.10.65",
-    "web-vitals": "^3.4.0",
     "xml2js": "^0.6.2"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "compilerOptions": {
-    "allowJs": true,
-    "outDir": "./dist"
-  },
   "devDependencies": {
-    "serverless-domain-manager": "^7.1.1",
-    "serverless-offline": "^12.0.4",
-    "serverless-webpack": "^5.13.0",
-    "webpack": "^5.88.2",
-    "webpack-node-externals": "^3.0.0"
+    "@astrojs/cloudflare": "^10.2.5",
+    "@astrojs/netlify": "^5.2.0",
+    "@astrojs/node": "^8.2.5",
+    "@astrojs/partytown": "^2.1.0",
+    "@astrojs/sitemap": "^3.1.4",
+    "@astrojs/svelte": "^5.4.0",
+    "@astrojs/ts-plugin": "^1.6.1",
+    "@astrojs/vercel": "^7.5.4",
+    "concurrently": "^8.2.2",
+    "nodemon": "^3.1.0",
+    "sass": "^1.77.1"
   }
 }
